backend/requirements.txt
fastapi
uvicorn
crewai
langchain_groq
langchain_google_genai
python-dotenv
litellm
crewai-tools
httpx


pypdf
langchain-community
langchain-google-genai
supabase